+++ python/sample_files/func_multi_var.py
@@ -1,3 +1,17 @@
+class A():
+    def __init__(self):
+        x = 1
+    
+    def sign(self, y):
+        if y < 0:
+            sign = "-"
+        elif y == 0:
+            sign = ""
+        else:
+            sign = "+"
+        
+        return sign
+
 def fun(a):
     if a:
         x = 1